% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greet_yell.R
\name{greet_yell}
\alias{greet_yell}
\title{Say Hey Loudly}
\usage{
greet_yell(name)
}
\arguments{
\item{name}{Name of user}
}
\value{
A string, all caps
}
\description{
Say Hey Loudly
}
\examples{
greet_yell('caro')
}
